Refactor: Add MemoryLayout class to help interleave the vertex data (WIP)
For now, only OBJ files are supported.
The data is now stored in separate arrays for clarity, and only interleaved during the creation of the GL buffer.
This should help us later in using heterogeneous data (non-floats).

It also makes setting the offsets in the rendering plugins more clear, instead of hard-coding offsets there.